# Conflicts:
#	src/main/resources/templates/layout/project/info.html

+++ src/main/java/com/project/ddbb/controller/ProjectMemberController.java
@@ -2,19 +2,11 @@ package com.project.ddbb.controller;
 
 import com.project.ddbb.domain.service.MemberService;
 import com.project.ddbb.domain.service.ProjectMemberService;
-import com.project.ddbb.domain.vo.MemberVO;
 import com.project.ddbb.domain.vo.ProjectMemberVO;
-import com.project.ddbb.domain.vo.ProjectVO;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-
-import java.util.List;
-import java.util.Map;
 
 @Controller
 @RequiredArgsConstructor
@@ -26,17 +18,18 @@ public class ProjectMemberController {
 
     /**
      * 프로젝트 참여자 등록
-     * @param id
+     * @param addId
      */
     @PostMapping("/add")
-    public String addProjectMember(RedirectAttributes redirect, @RequestParam String id, @RequestParam Long projectId,@RequestParam int leaderYn) throws Exception {
-        if (memberService.accountPermitId(id)) {
+    public String addProjectMember(RedirectAttributes redirect, @RequestParam String addId, @RequestParam Long projectId) throws Exception {
+        if (memberService.accountPermitId(addId)) {
             ProjectMemberVO pmv = new ProjectMemberVO();
-            pmv.setMemberId(memberService.selectById(id).getMemberId());
+            pmv.setMemberId(memberService.selectById(addId).getMemberId());
             pmv.setProjectId(projectId);
-            pmv.setLeaderYn(leaderYn);
+            pmv.setLeaderYn(false);
             projectMemberService.save(pmv);
         } else {
+            redirect.addAttribute("projectId", projectId);
             return "redirect:/project/info";
         }
         redirect.addAttribute("projectId", projectId);
@@ -45,11 +38,16 @@ public class ProjectMemberController {
 
     /**
      * 프로젝트 참여자 삭제
-     * @param vo
+     * @param deleteId
      */
     @PostMapping("delete")
-    @ResponseBody
-    public void deleteProjectMember(@RequestParam("ProjectMemberVO") ProjectMemberVO vo) {
-        projectMemberService.delete(vo);
+    public String deleteProjectMember(RedirectAttributes redirect, @RequestParam String deleteId, @RequestParam Long projectId) throws Exception {
+        Long memId = memberService.selectById(deleteId).getMemberId();
+        ProjectMemberVO vo = projectMemberService.findByProjectMember(memId, projectId);
+        if(vo != null)
+            projectMemberService.delete(vo);
+
+        redirect.addAttribute("projectId", projectId);
+        return "redirect:/project/info";
     }
 }
